incr: remove src imports from ipcCalls.ts - simplify ipcCalls
