+++ reports/commonExporter.ts
@@ -1,11 +1,12 @@
 import { t } from 'fyo';
 import { Action } from 'fyo/model/types';
 import { Verb } from 'fyo/telemetry/types';
-import { getSavePath, saveData, showExportInFolder } from 'src/utils/ipcCalls';
+import { saveData } from 'src/utils/ipcCalls';
 import { getIsNullOrUndef } from 'utils';
 import { generateCSV } from 'utils/csvParser';
 import { Report } from './Report';
 import { ExportExtention, ReportCell } from './types';
+import { getSavePath, showExportInFolder } from 'src/utils/ui';
 
 interface JSONExport {
   columns: { fieldname: string; label: string }[];
