+++ src/pages/TemplateBuilder/TemplateBuilder.vue
@@ -233,23 +233,23 @@ import ShortcutKeys from 'src/components/ShortcutKeys.vue';
 import { handleErrorWithDialog } from 'src/errorHandling';
 import { shortcutsKey } from 'src/utils/injectionKeys';
 import { showDialog, showToast } from 'src/utils/interactive';
-import { getSavePath } from 'src/utils/ipcCalls';
 import { docsPathMap } from 'src/utils/misc';
 import {
-  PrintTemplateHint,
-  baseTemplate,
-  getPrintTemplatePropHints,
-  getPrintTemplatePropValues,
+PrintTemplateHint,
+baseTemplate,
+getPrintTemplatePropHints,
+getPrintTemplatePropValues,
 } from 'src/utils/printTemplates';
 import { docsPathRef, showSidebar } from 'src/utils/refs';
 import { DocRef, PrintValues } from 'src/utils/types';
 import {
-  ShortcutKey,
-  focusOrSelectFormControl,
-  getActionsForDoc,
-  getDocFromNameIfExistsElseNew,
-  openSettings,
-  selectTextFile,
+ShortcutKey,
+focusOrSelectFormControl,
+getActionsForDoc,
+getDocFromNameIfExistsElseNew,
+getSavePath,
+openSettings,
+selectTextFile,
 } from 'src/utils/ui';
 import { useDocShortcuts } from 'src/utils/vueUtils';
 import { getMapFromList } from 'utils/index';
