+++ src/utils/ipcCalls.ts
@@ -2,8 +2,12 @@
  * Utils that make ipcRenderer calls.
  */
 const { ipcRenderer } = require('electron');
-import { t } from 'fyo';
-import { BaseError } from 'fyo/utils/errors';
+import type {
+  OpenDialogOptions,
+  OpenDialogReturnValue,
+  SaveDialogOptions,
+  SaveDialogReturnValue,
+} from 'electron';
 import type { BackendResponse } from 'utils/ipc/types';
 import { IPC_ACTIONS, IPC_CHANNELS, IPC_MESSAGES } from 'utils/messages';
 import type {
@@ -13,9 +17,6 @@ import type {
   SelectFileReturn,
   TemplateFile,
 } from 'utils/types';
-import { showDialog, showToast } from './interactive';
-import { setLanguageMap } from './language';
-import type { OpenDialogReturnValue } from 'electron';
 
 export function reloadWindow() {
   return ipcRenderer.send(IPC_MESSAGES.RELOAD_MAIN_WINDOW);
@@ -29,12 +30,11 @@ export async function getLanguageMap(code: string) {
   };
 }
 
-export async function getSelectedFilePath(): Promise<OpenDialogReturnValue> {
-  return (await ipcRenderer.invoke(IPC_ACTIONS.GET_OPEN_FILEPATH, {
-    title: t`Select file`,
-    properties: ['openFile'],
-    filters: [{ name: 'SQLite DB File', extensions: ['db'] }],
-  })) as OpenDialogReturnValue;
+export async function getOpenFilePath(options: OpenDialogOptions) {
+  return (await ipcRenderer.invoke(
+    IPC_ACTIONS.GET_OPEN_FILEPATH,
+    options
+  )) as OpenDialogReturnValue;
 }
 
 export async function getTemplates(): Promise<TemplateFile[]> {
@@ -61,43 +61,17 @@ export async function checkDbAccess(filePath: string) {
 
 export async function checkForUpdates() {
   await ipcRenderer.invoke(IPC_ACTIONS.CHECK_FOR_UPDATES);
-  await setLanguageMap();
 }
 
 export function openLink(link: string) {
   ipcRenderer.send(IPC_MESSAGES.OPEN_EXTERNAL, link);
 }
 
-export async function deleteDb(filePath: string) {
-  const { error } = (await ipcRenderer.invoke(
+export async function deleteFile(filePath: string) {
+  return (await ipcRenderer.invoke(
     IPC_ACTIONS.DELETE_FILE,
     filePath
   )) as BackendResponse;
-
-  if (error?.code === 'EBUSY') {
-    await showDialog({
-      title: t`Delete Failed`,
-      detail: t`Please restart and try again.`,
-      type: 'error',
-    });
-  } else if (error?.code === 'ENOENT') {
-    await showDialog({
-      title: t`Delete Failed`,
-      detail: t`File ${filePath} does not exist.`,
-      type: 'error',
-    });
-  } else if (error?.code === 'EPERM') {
-    await showDialog({
-      title: t`Cannot Delete`,
-      detail: t`Close Frappe Books and try manually.`,
-      type: 'error',
-    });
-  } else if (error) {
-    const err = new BaseError(500, error.message);
-    err.name = error.name;
-    err.stack = error.stack;
-    throw err;
-  }
 }
 
 export async function saveData(data: string, savePath: string) {
@@ -113,47 +87,21 @@ export async function makePDF(
   savePath: string,
   width: number,
   height: number
-): Promise<void> {
-  const success = (await ipcRenderer.invoke(
+): Promise<boolean> {
+  return (await ipcRenderer.invoke(
     IPC_ACTIONS.SAVE_HTML_AS_PDF,
     html,
     savePath,
     width,
     height
   )) as boolean;
-
-  if (success) {
-    showExportInFolder(t`Save as PDF Successful`, savePath);
-  } else {
-    showToast({ message: t`Export Failed`, type: 'error' });
-  }
 }
 
-export function showExportInFolder(message: string, filePath: string) {
-  showToast({
-    message,
-    actionText: t`Open Folder`,
-    type: 'success',
-    action: () => {
-      showItemInFolder(filePath);
-    },
-  });
-}
-
-export async function getSavePath(name: string, extention: string) {
-  const response = (await ipcRenderer.invoke(IPC_ACTIONS.GET_SAVE_FILEPATH, {
-    title: t`Select Folder`,
-    defaultPath: `${name}.${extention}`,
-  })) as { canceled: boolean; filePath?: string };
-
-  const canceled = response.canceled;
-  let filePath = response.filePath;
-
-  if (filePath && !filePath.endsWith(extention) && filePath !== ':memory:') {
-    filePath = `${filePath}.${extention}`;
-  }
-
-  return { canceled, filePath };
+export async function getSaveFilePath(options: SaveDialogOptions) {
+  return (await ipcRenderer.invoke(
+    IPC_ACTIONS.GET_SAVE_FILEPATH,
+    options
+  )) as SaveDialogReturnValue;
 }
 
 export async function getDbList() {
